create generic compose player dialog
